requirements.txt
astrapy==2.1.0
clip_anytorch==2.6.0
fastapi==0.121.0
langchain==1.0.3
langchain_astradb==1.0.0
langchain_community==0.4.1
langchain_core==1.0.3
langchain_unstructured==0.1.6
numpy==2.3.4
openpyxl==3.1.5
pandas==2.3.3
Pillow==12.0.0
python-dotenv==1.2.1
python_docx==1.2.0
python_pptx==0.6.23
Requests==2.32.5
streamlit==1.49.1
torch==2.8.0+cu126
unstructured==0.10.30
utility==1.0
#pipreqs . --force --savepath=requirements.txt
